requirements.txt
Flask
Telethon
firebase-admin
Flask-Cors
python-dotenv